requirements.txt
asyncio
aiohttp
pillow